chore(tailwind): migrate tailwind config to css
* tailwind config is still required for typography
