+++ tailwind.config.ts
@@ -1,0 +1,49 @@
+import type { Config } from "tailwindcss";
+
+const config = {
+  theme: {
+    extend: {
+      typography: () => ({
+        DEFAULT: {
+          css: [
+            {
+              "--tw-prose-body": "var(--text)",
+              "--tw-prose-headings": "var(--text)",
+              h1: {
+                fontWeight: "normal",
+                marginBottom: "0.25em",
+              },
+            },
+          ],
+        },
+        base: {
+          css: [
+            {
+              h1: {
+                fontSize: "2.5rem",
+              },
+              h2: {
+                fontSize: "1.25rem",
+                fontWeight: 600,
+              },
+            },
+          ],
+        },
+        md: {
+          css: [
+            {
+              h1: {
+                fontSize: "3.5rem",
+              },
+              h2: {
+                fontSize: "1.5rem",
+              },
+            },
+          ],
+        },
+      }),
+    },
+  },
+} satisfies Config;
+
+export default config;
